requirements.txt
django
django-multiselectfield
pandas
numpy
django-widgets-improved
scikit-learn
joblib
lightgbm
python-dotenv